requirements.txt
altair
pandas
streamlit
numpy
tensorflow
pillow
werkzeug
cloudinary